requirements.txt
streamlit==1.28.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0
langchain==0.1.20
langchain-google-genai==2.0.0
sqlalchemy==2.0.23
sentence-transformers==2.2.2
